requirements.txt
google-api-core==1.26.3
google-api-python-client==2.23.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
gunicorn==20.0.4
json5==0.9.5
jsonschema==3.2.0
numpy==1.18.5
oauth2client==3.0.0
oauthlib==3.1.0
pandas==1.1.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
uvicorn==0.13.4